requirements.txt
flake8==6.0.0
Flask==2.2.2
pandas==1.5.2
psycopg2==2.9.5
python-dotenv==0.21.0
xlrd==2.0.1
